refactor. delete junk
