--- Atom_Projected_Phonon_Dispersion/README.md
@@ -1,25 +1,0 @@
-# Atom-Projected Phonon Dispersion Plotter
-
-This Python script reads phonon dispersion data from a phonopy output file "band.yaml" and plots the atom-projected phonon dispersion using the corresponding eigenvectors. The script provides options to generate scatter plots and colormap (cmap) plots for visualizing the atom contributions to the phonon modes.
-
-## Prerequisites
-
-1. Ensure that the "band.yaml" file is present in the working directory, containing the phonon dispersion data as generated by the phonopy package.
-2. This version best works for simple diatomic structures.
-
-## Features
-
-1. Load phonon dispersion data from the "band.yaml" file.
-2. Extract eigenvectors representing atom-projected contributions.
-3. Normalize amplitudes of eigenvectors for visualization.
-4. Plot atom-projected phonon dispersion using scatter plots or colormap plots.
-
-## Usage
-
-1. Install the required dependencies: `numpy`, `matplotlib`.
-2. Place the "band.yaml" file in the working directory. YbS band.yaml file is supplied with the code.
-3. Run the script using the following command:
-
-```shell
-python your_script.py [--scatter] [--cmap COLORMAP] [--gnuplot] [--atom1_label LABEL] [--atom2_label LABEL]
-
