+++ README.md
@@ -1,5 +1,6 @@
 # PLATO: Phonon LAttice TOols a place for all lattice dynamics and thermal transport related tools.
 
-1. Tdep_helping_Tools my set of tools I wrote few years back to work with TDEP output files. Check README.md inside directory to find out more.
+1. Tdep_Helping_Tools my set of tools I wrote few years back to work with TDEP output files. Check README.md inside directory to find out more.
  
-2. Atom_Projected_Phonon_Dispersion provies a way to resolve the phonon dispersion in terms of atomic contributions.
+2. Phonopy_Phono3py__Helping_Tools is set to tools to work with phono(3)py output. It provides ways for quick and better data analysis.
+
